+++ $P@/SPA/SPA/Helper.cpp
@@ -34,7 +34,7 @@ string Helper::charToString(char c)
 int Helper::stringToInt(string s)
 {
 	int ans = 0;
-	for (unsigned int i = 0; i < s.size(); i++)
+	for (size_t i = 0; i < s.size(); i++)
 		ans = ans * 10 + (s[i] - '0');
 	return ans;
 }
